# src/swagger/swaggerDocumentation.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "1.0.0",
    "title": "com3504",
    "description": "User management API",
    "termsOfService": "http://api_url/terms/",
    "contact": {
      "name": "Boxuan Shan, Jingxiang Qi, Ziting Zhao",
      "email": "bshan3@sheffield.ac.uk",
      "url": "https://github.com/ace18zz/COM3504The-Intelligent-Web"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "the only Local server"
    }
  ],
  "paths": {
...
    "/search": {
      "get": {
          "description": " Search room by auther",
          "parameters": [
              {
                  "name": "authorName",
                  "in": "query",
                  "required": true,
                  "schema": {
                      "type": "string"
                  },
                  "allowEmptyValue": true
              }
          ],
          "operationId": "searchAuthor",
          "responses": {
              "default": {
                  "description": "Correct input of search structure",
                  "content": {
                      "application/json": {
                        "schema": {
                          "type": "array",
                          "items":{
                          "$ref": "#/components/schemas/ImgMetaObj"
                          }

                      }
                  } 
              }
          }
      }
  }
},
...
},
"components": {
  "schemas": {
    "statusCode": {
      "type": "integer",
      "description": "Internal status code for response, value 0 for no error",
      "example": 0
    },
    "imgId": {
      "type": "string",
      "description": "Image ID",
      "example": "605c6c684431fa7cd4b36134"
    },
    "imgTitle": {
      "type": "string",
      "description": "Image title",
      "example": "Untitled - 1"
    },
    "imgAuthor": {
      "type": "string",
      "description": "Image author",
      "example": "User1"
    },
    "imgDesc": {
      "type": "string",
      "description": "Image description",
      "example": "This is an image"
    },
...
    "ImgMetaObj": {
      "type": "object",
      "properties": {
        "code":{
          "$ref":"#/components/schemas/statusCode"
        },
        "id": {
          "$ref": "#/components/schemas/imgId"
        },
        "title": {
          "$ref": "#/components/schemas/imgTitle"
        },
        "author": {
          "$ref": "#/components/schemas/imgAuthor"
        },
        "desc": {
          "$ref": "#/components/schemas/imgDesc"
        }
      }
    },
...
  }
}
}